requirements.txt
google-genai>=1.0.0
pdf2image>=1.16.0
Pillow>=10.0.0
tqdm>=4.60.0
python-dotenv>=1.0.0
python-docx>=1.1.0
markdown>=3.5.0
